requirements.txt
flask
python-dotenv
openai
PyMuPDF
python-docx
